♻️ :: VoteResponse 분리
