--- src/main/java/com/project/dmsport/domain/club/presentation/response/QueryVoteHistoryListResponse.java
@@ -1,40 +1,0 @@
-package com.project.dmsport.domain.club.presentation.response;
-
-import com.project.dmsport.domain.club.domain.Club;
-import com.project.dmsport.domain.vote.domain.Vote;
-import com.project.dmsport.domain.vote.domain.enums.VoteType;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-import java.time.LocalDate;
-import java.util.List;
-import java.util.Map;
-
-@Getter
-@AllArgsConstructor
-public class QueryVoteHistoryListResponse {
-
-        private final Map<LocalDate, List<VoteResponse>> list;
-
-        @Getter
-        @Builder
-        public static class VoteResponse {
-                private final Long voteId;
-                private final VoteType time;
-                private final Integer voteCount;
-                private final Integer maxPeople;
-                private final List<String> voteUserName;
-        }
-
-        public static VoteResponse of(Vote vote, Club club, List<String> voteUserList) {
-                return VoteResponse.builder()
-                        .voteId(vote.getId())
-                        .time(vote.getVoteType())
-                        .voteCount(vote.getCount())
-                        .maxPeople(club.getMaxPeople())
-                        .voteUserName(voteUserList)
-                        .build();
-        }
-
-}
