+++ src/main/java/com/project/dmsport/domain/vote/presentation/dto/response/QueryTodayVoteListResponse.java
@@ -1,4 +1,4 @@
-package com.project.dmsport.domain.club.presentation.response;
+package com.project.dmsport.domain.vote.presentation.dto.response;
 
 import com.project.dmsport.domain.vote.domain.enums.VoteType;
 import lombok.Builder;
@@ -14,14 +14,6 @@ public class QueryTodayVoteListResponse {
     private final boolean isBan;
     private final LocalDate banPeriod;
     private final Integer maxPeople;
-    private final List<VoteResponse> vote;
-
-    @Getter
-    @Builder
-    public static class VoteResponse {
-        private final Long voteId;
-        private final VoteType time;
-        private final Integer voteCount;
-    }
+    private final List<VoteResponse> voteList;
 
 }
